Move validateXml from test-utils
