+++ forms-shared/tests/slovensko-sk/slovenskoSkForm.ts
@@ -9,7 +9,7 @@ import { getHtmlSbXslt } from '../../src/slovensko-sk/file-templates/htmlSbXslt'
 import { generatePageScreenshot } from '../../test-utils/generatePageScreenshot'
 import { getSchemaXsd } from '../../src/slovensko-sk/file-templates/schemaXsd'
 import { formDefinitions } from '../../src/definitions/formDefinitions'
-import { validateXml } from '../../test-utils/validateXml'
+import { validateXml } from '../../src/slovensko-sk/validateXml'
 import { fetchSlovenskoSkFormMetadata } from '../../test-utils/fetchSlovenskoSkFormMetadata'
 import { extractJsonFromSlovenskoSkXml } from '../../src/slovensko-sk/extractJson'
 import { buildSlovenskoSkXml } from '../../src/slovensko-sk/xmlBuilder'
